move EntityExistsValidator class into separate package, refactor exception message
